pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-platform-monorepo"
version = "0.2.0"
description = "Monorepo for knowledge-centric RAG platform (API, core packages, future web)"
authors = [{ name = "Your Name" }]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0,<0.112.0",
    "uvicorn[standard]>=0.30.0,<0.31.0",
    "langchain>=0.2.4,<0.3.0",
    "langchain-core>=0.2.4,<0.3.0",
    "langgraph>=0.0.56,<0.1.0",
    "langchain-community>=0.2.4,<0.3.0",
    "faiss-cpu>=1.8.0",
    "pydantic>=2.7.0,<3.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "pypdf>=4.0.0,<5.0.0",
    "python-docx>=1.1.0,<2.0.0",
    "python-multipart>=0.0.9,<0.1.0",
    "langchain-openai>=0.1.25",
    "openai>=1.0.0",
    "tenacity>=8.0.0",
    "httpx[socks]>=0.27.0",
    "socksio>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
    "httpx>=0.27.0,<0.28.0",
    "ruff>=0.5.0,<0.6.0",
]

[tool.setuptools.packages.find]
where = ["apps/api", "packages/rag-core", "packages/shared-config", "packages/domain-models"]
include = ["*"]
exclude = [
    "*.tests",
    "tests",
    "tests.*",
    "*.migrations",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
    "httpx>=0.27.0,<0.28.0",
    "ruff>=0.5.0,<0.6.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["rag_core", "shared_config", "domain_models"]
